flaskweb/requirements.txt
awscli==1.6.2
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bcdoc==0.12.2
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
binaryornot==0.4.0
boto==2.33.0
botocore==0.73.0
certifi==2015.4.28
cffi==1.4.2
characteristic==14.3.0
chardet==2.3.0
click==6.2
colorama==0.2.5
cookiecutter==1.3.0
cookiejar==0.0.1
coverage==4.0.3
cryptography==1.1.2
cssselect==0.9.1
cycler==0.9.0
decorator==3.4.0
Django==1.8.5
docutils==0.12
dominate==2.2.0
enum34==1.1.2
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
functools32==3.2.3.post2
future==0.15.2
gunicorn==19.4.5
idna==2.0
ipaddress==1.0.16
ipykernel==4.2.1
ipython==4.0.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
Jinja2==2.8
jmespath==0.5.0
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
librato-metrics==0.4.9
lxml==3.5.0
MarkupSafe==0.23
mistune==0.7.1
mixpanel-py==3.1.1
multipledispatch==0.4.8
mysql==0.0.1
MySQL-python==1.2.3
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.0.6
numpy==1.10.1
pager==3.3
path.py==8.1.2
pbr==1.8.1
pickleshare==0.5
prettytable==0.7.2
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
Pygments==2.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.6
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
pyzmq==15.1.0
qtconsole==4.1.1
queuelib==1.4.2
requests==2.3.0
rsa==3.1.2
Scrapy==1.0.4
service-identity==14.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.0.12
sseclient==0.0.8
stevedore==1.12.0
tornado==4.1
tqdm==2.0.0
traitlets==4.0.0
Twisted==15.5.0
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
visitor==0.1.2
w3lib==1.13.0
web.py==0.37
Werkzeug==0.11.5
whichcraft==0.1.1
WTForms==2.1
zope.interface==4.1.3
